pyproject.toml
[project]
name = "kira"
version = "0.1.0"
description = "KIRA (Krafton Intelligence Rookie Agent)"
authors = [{name = "KRAFTON AI"}]
readme = "README.md"
requires-python = ">=3.10,<4.0"

dependencies = [
  "google-adk>=1.10.0",
  "pydantic>=2.11.7",
  "slack_bolt>=1.18.0",
  "python-dotenv>=1.0.0",
  "apscheduler>=3.10.4",
  "python-dateutil>=2.8.2",
  "litellm>=1.75.5.post1",
  "a2a-sdk>=0.3.1",
  "arize-phoenix-otel>=0.13.0",
  "openinference-instrumentation-google-adk>=0.1.3",
  "watchdog>=6.0.0",
  "atlassian-python-api>=4.0.7",
  "psutil>=7.1.0",
  "pandas>=2.3.3",
  "openpyxl>=3.1.5",
  "pypdf>=6.1.1",
  "termcolor>=3.1.0",
  "playwright>=1.52.0",
  "claude-agent-sdk>=0.1.4",
  "anthropic[vertex]>=0.71.0",
  "python-pptx>=1.0.2",
  "typing-extensions>=4.12.2",
  "msal>=1.34.0",
  "fastapi>=0.119.0",
  "uvicorn>=0.38.0",
  "websockets>=15.0.1",
  "python-multipart>=0.0.20",
  "authlib>=1.6.5",
  "itsdangerous>=2.2.0",
  "httpx>=0.28.1",
  "tweepy>=4.16.0",
  "requests>=2.32.0",
  "requests-oauthlib>=2.0.0",
  "imageio>=2.37.0",
  "numpy>=2.2.6",
  "markdown>=3.9",
  "reportlab>=4.4.4",
  "beautifulsoup4>=4.14.2",
  "matplotlib>=3.10.7",
  "markdown2>=2.5.4",
  "weasyprint>=66.0",
  "python-docx>=1.2.0",
  "pip-licenses>=5.5.0",
]

[dependency-groups]
dev = [
    "black>=23.0.0",
    "devtools>=0.12.2",
    "flask>=2.3.3",
]

[tool.black]
target-version = ['py310']
